Order the node interface functions
